ubuntu 24 - build support
The core issue was the need for `llvm-11` source compilation with the correct options/features enabled.  To do so, required `gcc-10`, which may need to be built from source itself in the near future, once installed and CMAKE was configured to use it as the compiler and with an install prefix path configured:

```sh
cmake \
-DCMAKE_C_COMPILER=gcc-10 \
-DCMAKE_CXX_COMPILER=g++-10 \
-DCMAKE_INSTALL_PREFIX=/opt/llvm/llvm-11 \
-DCMAKE_PREFIX_PATH=/opt/llvm/llvm-11 \
...
```
The scripts for building llvm properly are under `scripts/llvm-11`.

The Dockerfile was also updated to build on Ubuntu 24.04. Additionally it was made multi-stage for build performance and efficency